working in skills routes

+++ controllers/aboutControllers.js
@@ -58,38 +58,42 @@ export const updateAbout = async (req, res, next) => {
     const id = req.params.id;
     const { fullName, jobTitle, description, mediaLinks } = req.body;
     const oldData = await About.findById(id);
-
-    let fileName;
-    if (req.file) {
-      let oldFilePath = oldData.profilePicture;
-      let localHostUrl = "http://localhost:8000/".length;
-      let path = oldFilePath.slice(localHostUrl);
-      fs.unlink(path, (err) => {
-        if (err) {
-          console.log("failed");
-        } else {
-          console.log("delete");
-        }
-      });
-      fileName = `${process.env.https + req.file.path.replaceAll("\\", "/")}`;
-      console.log(fileName);
-    }
-    const update = await About.findByIdAndUpdate(
-      id,
-      {
-        fullName,
-        jobTitle,
-        description,
-        mediaLinks,
-        profilePicture: fileName,
-      },
-      {
-        new: true,
+    if (oldData) {
+      let fileName;
+      if (req.file) {
+        let oldFilePath = oldData.profilePicture;
+        let localHostUrl = "http://localhost:8000/".length;
+        let path = oldFilePath.slice(localHostUrl);
+        fs.unlink(path, (err) => {
+          if (err) {
+            console.log("failed");
+          } else {
+            console.log("delete");
+          }
+        });
+        fileName = `${process.env.https + req.file.path.replaceAll("\\", "/")}`;
+        // console.log(fileName);
       }
-    );
-    return res.status(201).json({
-      message: "Sucessful",
-      update,
+      const update = await About.findByIdAndUpdate(
+        id,
+        {
+          fullName,
+          jobTitle,
+          description,
+          mediaLinks,
+          profilePicture: fileName,
+        },
+        {
+          new: true,
+        }
+      );
+      return res.status(201).json({
+        message: "Sucessful",
+        update,
+      });
+    }
+    return res.status(404).json({
+      message: "id not found ",
     });
   } catch (error) {
     return res.status(400).json({
@@ -98,4 +102,37 @@ export const updateAbout = async (req, res, next) => {
   }
 };
 
-export const deleteAbout = (req, res, next) => {};
+export const deleteAbout = async (req, res, next) => {
+  try {
+    const id = req.params.id;
+    const oldData = await About.findById(id);
+    // console.log(oldData);
+    if (!oldData) {
+      return res.status(404).json({
+        message: "Id not found",
+      });
+    }
+    await About.findByIdAndDelete(id);
+    if (oldData) {
+      let imagePath = oldData.profilePicture.replace(
+        "http://localhost:8000/",
+        ""
+      );
+      // console.log(imagePath);
+      fs.unlink(imagePath, (err) => {
+        if (err) {
+          console.log("failed", err);
+        } else {
+          console.log("sucess");
+        }
+      });
+    }
+    return res.status(201).json({
+      message: "passed",
+    });
+  } catch (error) {
+    return res.status(400).json({
+      message: "failed",
+    });
+  }
+};
